Add info for Rallyman DIRT 110% expansion tiles
Based on tile preview images posted in Kickstarter updates. Some tile
numbers are guessed when they are not shown in the image.

Also the filler tile shown in update #15 is assumed to be part of the
110% expansion, even though this is not stated explicitly in the update post
text. It makes sense from the tile art and the total count of expansion
tiles.

Part of ticket #10.