Adjust map size when loading large tracks
Checks whether the current map is larger than the default size and
increases the map area accordingly. Shrinks back to default size in case
tiles are removed.

This approach makes all tiles accessible for large track files, like
some of the Hill Climb tracks. It does not provide full map area
management with pan and zoom controls. This is planned for a later
release (see issue #9).