Add watchdog to check if WASM initialization was successful
At the beginning of the year 2022 the Rekee website was broken for
visitors using Chrome browser updated to version 100. As the page
continued to work in Firefox it took some months till the author noticed
the breakage.

As an alternative to the usage of an third-party website tracking
service this implementation uses a JavaScript timer to check whether
catalog tiles and the map have been initialized correctly by the
WebAssembly module. If the check fails a warning message is shown with
instructions where to report the issue.