Use separate group element for map tile labels
Prepare for token display on the map view. As SVG 1.1 does not support a
"z-index" attribute the draw order must be specified via the element
order. By using separate group elements for the tile image and labels we
can insert a group element for tokens later between them.

Also invert the visible/hidden logic of catalog tiles for symmetry with
the map logic.