Add token support to image export
Use a separate list of token images that are loaded and drawn after tile
images are complete. Put drawing of tile labels at the end of the
processing when all images are done.

Unify image size and center calculations for tiles and tokens used in
catalog, map, and image export.

Implements issues #12 and #13 for the image export component.