Use larger tile thumbnail images
Thumbnail image size of tiles is enlarged from 173x150 to 240x208
pixels, this is double the default display size and should provide more
crispy images when scaling the webpage in the browser. Image format is
changed from PNG to WebP to keep the file size, with similar optical
image quality.

Always use the "lower hex" string formatting when generating file URLs
for consistency.