Allow to add and delete tile tokens
Implement card footer actions that allow to create and delete a token.
The token image properties card object is changed to remove its HTML
elements automatically when dropped. The index number is hidden in card
titles as it will not be sequential anymore after some mixed add and
delete events.