Detect whether the browser Canvas API supports font features
Not all current browsers seem to support the configuration of font
features as done in 76fa2802132ac956a916e4e0c2dda3a77b4af638, for
example it fails in Chrome v108 or Edge v107 on Windows.

Implement a work-around by testing the font after it has been loaded
regarding the measured width of numeric glyphs and load a replacement
font when necessary.

The replacement font file is a patched variant of Overpass v4.0 where
the default Unicode entries for digit zero to nine have been changed to
the '.tf' glyph variants.