Replace the default "My Track" map title with an empty string
Found some track designs on the web where the page context did mention a
specific track title but the embedded image still displayed the generic
"My Track" header. Apparently it makes more sense to let the users
specify a title when actually needed, instead of forcing them to clear
the string when not needed.