Support tile listing box on image export
Allows to draw a tile listing summary text box on image export. The
style is similar to listing boxes used in the track examples of Rallyman
DIRT rulebooks with the limitation that only a single column is
rendered, and the "times" operator uses the corresponding mathematical
glyph instead of a "x" letter.