Optimizations for the build procedure
Always set the "cargo:rerun-if" flags in the build script. Otherwise it
will not work correctly when switching between feature flags. Note that
setting the flag for build.rs itself is redundant, according to Cargo
documentation.

Include a second WASM optimization pass when running make with the
"build-release" target and when "wasm-opt" is available system-wide. The
second optimization run shrinks the output file from 528k to 520k.